+++ src/canvas/Canvas.spec.ts
@@ -1,5 +1,7 @@
 import { Canvas } from './Canvas';
 import { Rect } from '../shapes/Rect';
+import { IText } from '../shapes/IText/IText';
+import '../shapes/ActiveSelection';
 
 describe('Canvas', () => {
   describe('touchStart', () => {
@@ -109,6 +111,61 @@ describe('Canvas', () => {
       evt.preventDefault = jest.fn();
       canvas._onTouchStart(evt);
       expect(evt.preventDefault).not.toHaveBeenCalled();
+    });
+    test('dispose after _onTouchStart', () => {
+      jest.spyOn(global, 'clearTimeout');
+      const canvas = new Canvas(undefined, {
+        allowTouchScrolling: true,
+        isDrawingMode: true,
+      });
+      const touch = {
+        clientX: 10,
+        clientY: 0,
+        identifier: 1,
+        target: canvas.upperCanvasEl,
+      };
+      const evtStart = new TouchEvent('touchstart', {
+        touches: [touch],
+        changedTouches: [touch],
+      });
+      canvas._onTouchStart(evtStart);
+      const evtEnd = new TouchEvent('touchend', {
+        touches: [],
+        changedTouches: [touch],
+      });
+      canvas._onTouchEnd(evtEnd);
+      expect(canvas._willAddMouseDown).toBeGreaterThan(0);
+      canvas.dispose();
+      expect(global.clearTimeout).toHaveBeenCalledWith(
+        canvas._willAddMouseDown,
+      );
+    });
+  });
+
+  describe('handleMultiSelection', () => {
+    const canvas = new Canvas();
+    const rect = new Rect({ left: 100, width: 100, height: 100 });
+    const iText = new IText('itext');
+    canvas.add(rect, iText);
+    test('Selecting shapes containing text does not trigger the exit event', () => {
+      const exitMock = jest.fn();
+      iText.on('editing:exited', exitMock);
+
+      const firstClick = new MouseEvent('click', {
+        clientX: 0,
+        clientY: 0,
+      });
+      canvas._onMouseDown(firstClick);
+      canvas._onMouseUp(firstClick);
+      const secondClick = new MouseEvent('click', {
+        shiftKey: true,
+        clientX: 100,
+        clientY: 0,
+      });
+      canvas._onMouseDown(secondClick);
+      canvas._onMouseUp(secondClick);
+
+      expect(exitMock).toHaveBeenCalledTimes(0);
     });
   });
 });