+++ src/canvas/Canvas.ts
@@ -198,7 +198,7 @@ export class Canvas extends SelectableCanvas implements CanvasOptions {
   }
 
   /**
-   * Removes all event listeners
+   * Removes all event listeners, used when disposing the instance
    */
   removeListeners() {
     this.addOrRemove(removeListener, 'remove');
@@ -228,6 +228,7 @@ export class Canvas extends SelectableCanvas implements CanvasOptions {
       this._onMouseMove as EventListener,
       addEventOptions,
     );
+    clearTimeout(this._willAddMouseDown);
   }
 
   /**
@@ -1445,7 +1446,7 @@ export class Canvas extends SelectableCanvas implements CanvasOptions {
         }
         this._fireSelectionEvents(prevActiveObjects, e);
       } else {
-        (activeObject as IText).exitEditing &&
+        (activeObject as IText).isEditing &&
           (activeObject as IText).exitEditing();
         // add the active object and the target to the active selection and set it as the active object
         const klass =