+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [next]
 
 - fix(TS): Add missing export for type DrawContext [#10281](https://github.com/fabricjs/fabric.js/pull/10281)
+- fix(Control): Borderscalefactor should apply to controls as well [#10283](https://github.com/fabricjs/fabric.js/pull/10283)
 
 ## [6.5.0]